# internal/api/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: fusion-index
  description: >
    Artifact registry — stores, indexes, and exposes versioned artifacts via REST.
    Artifacts use Python-style namespaced names (`org.team.name`), semver versions
    (`major.minor.patch`), per-version configurations, file uploads, and mutable tags.
  version: "0.1.0"
  license:
    name: GPL-3.0
    identifier: GPL-3.0-only
servers:
  - url: http://localhost:8080
    description: Local / port-forward
  - url: http://index.fusion.local
    description: Minikube ingress
...
components:
  parameters:
    artifactId:
      name: id
      in: path
      required: true
      description: Artifact ID
      schema:
        type: integer
        format: int64
    semver:
      name: semver
      in: path
      required: true
      description: Semantic version string (`major.minor.patch`)
      schema:
        type: string
        pattern: '^\d+\.\d+\.\d+$'
        example: "1.2.3"
    fileId:
      name: fileId
      in: path
      required: true
      description: File ID
      schema:
        type: integer
        format: int64
    tag:
      name: tag
      in: path
      required: true
      description: Tag name (e.g. `latest`, `stable`)
      schema:
        type: string
        maxLength: 255
    typeId:
      name: typeId
      in: path
      required: true
      description: Type ID
      schema:
        type: integer
        format: int64
    page:
      name: page
      in: query
      required: false
      description: Zero-based page number
      schema:
        type: integer
        minimum: 0
        default: 0
    pageSize:
      name: pageSize
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        minimum: 1
        default: 20
    olderThan:
      name: olderThan
      in: query
      required: true
      description: >
        Return only items created before this timestamp (RFC3339, e.g. `2024-01-01T00:00:00Z`).
      schema:
        type: string
        format: date-time
        example: "2024-01-01T00:00:00Z"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "artifact not found"
...
    TypeResponse:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
          example: "ml-model"
        description:
          type: ["string", "null"]
          example: "Machine learning model artifacts"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTypeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
          example: "ml-model"
        description:
          type: ["string", "null"]
          maxLength: 2000
          example: "Machine learning model artifacts"
    UpdateTypeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
          example: "ml-model"
        description:
          type: ["string", "null"]
          maxLength: 2000
    ArtifactResponse:
      type: object
      required: [id, fullName, types, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
          maxLength: 500
          example: "org.platform.data.pipeline"
        description:
          type: ["string", "null"]
          example: "ETL pipeline artifact"
        types:
          type: array
          items:
            $ref: "#/components/schemas/TypeResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArtifactPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
        pageSize:
          type: integer
    CreateArtifactRequest:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
          maxLength: 500
          example: "org.platform.data.pipeline"
        description:
          type: ["string", "null"]
          maxLength: 2000
          example: "ETL pipeline artifact"
    UpdateArtifactRequest:
      type: object
      properties:
        description:
          type: ["string", "null"]
          maxLength: 2000
    TagResponse:
      type: object
      required: [id, artifactId, tag, versionId, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        artifactId:
          type: integer
          format: int64
        tag:
          type: string
        versionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VersionResponse:
      type: object
      required: [id, artifactId, version, major, minor, patch, tags, createdAt]
      properties:
        id:
          type: integer
          format: int64
        artifactId:
          type: integer
          format: int64
        version:
          type: string
          example: "1.2.3"
        major:
          type: integer
          format: int32
        minor:
          type: integer
          format: int32
        patch:
          type: integer
          format: int32
        config:
          type: ["string", "null"]
          description: Raw configuration string (JSON or YAML)
          example: '{"workers": 4, "timeout": "30s"}'
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagResponse"
        createdAt:
          type: string
          format: date-time
    CreateVersionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          example: "1.2.3"
        config:
          type: ["string", "null"]
          description: Raw configuration string (JSON or YAML)
        tags:
          type: array
          items:
            type: string
          description: Tag names to assign to this version on creation
...
    AssignTagRequest:
      type: object
      required: [version]
      properties:
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          example: "1.2.3"
          description: Semver string of the version to assign this tag to
    FileResponse:
      type: object
      required: [id, versionId, name, storageBackend, status, downloadUrl, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        versionId:
          type: integer
          format: int64
        name:
          type: string
          example: "model.onnx"
        contentType:
          type: ["string", "null"]
          example: "application/octet-stream"
        sizeBytes:
          type: ["integer", "null"]
          format: int64
        storageBackend:
          type: string
          enum: [FILESYSTEM, S3]
        status:
          type: string
          enum: [PENDING, AVAILABLE, ERROR]
        downloadUrl:
          type: string
          example: "/api/v1/artifacts/1/versions/1.2.3/files/7/download"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/artifacts:
    get:
      tags: [Artifacts]
      summary: List artifacts
      description: >
        Returns a paginated list of artifacts.
        Use `name` for prefix search, `tag` to filter by assigned tag name, or `type`
        (repeatable) to filter by type name with OR semantics.
        Filter precedence: `type` > `name` > `tag`.
      operationId: listArtifacts
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - name: name
          in: query
          required: false
          description: Filter by full name prefix (e.g. `org.platform`)
          schema:
            type: string
        - name: tag
          in: query
          required: false
          description: Filter artifacts that have this tag assigned
          schema:
            type: string
        - name: type
          in: query
          required: false
          description: >
            Filter artifacts that have at least one of the given types assigned.
            Repeatable for OR semantics (e.g. `?type=ml-model&type=dataset`).
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Paginated artifact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactPage"

    post:
      tags: [Artifacts]
      summary: Create artifact
      operationId: createArtifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactRequest"
      responses:
        "201":
          description: Artifact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: An artifact with this full name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}:
    parameters:
      - $ref: "#/components/parameters/artifactId"

    get:
      tags: [Artifacts]
      summary: Get artifact
      operationId: getArtifact
      responses:
        "200":
          description: Artifact found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactResponse"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [Artifacts]
      summary: Update artifact description
      operationId: updateArtifact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtifactRequest"
      responses:
        "200":
          description: Artifact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Artifacts]
      summary: Delete artifact
      description: Deletes the artifact and all its versions, files, and tags (cascade).
      operationId: deleteArtifact
      responses:
        "204":
          description: Artifact deleted
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Versions --------------------------------------------------------------
  /api/v1/artifacts/{id}/versions:
    parameters:
      - $ref: "#/components/parameters/artifactId"

    get:
      tags: [Versions]
      summary: List versions
      description: Returns all versions for the artifact, ordered newest first (major, minor, patch DESC).
      operationId: listVersions
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VersionResponse"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Versions]
      summary: Create version
      description: >
        Creates a new version. Tags listed in `tags[]` are atomically assigned to
        this version (upsert semantics — a tag previously on another version moves here).
      operationId: createVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVersionRequest"
      responses:
        "201":
          description: Version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionResponse"
        "400":
          description: Invalid request body or malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Version already exists for this artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/versions/{semver}:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/semver"

    get:
      tags: [Versions]
      summary: Get version
      operationId: getVersion
      responses:
        "200":
          description: Version found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionResponse"
        "400":
          description: Malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Versions]
      summary: Delete version
      description: Deletes the version and performs best-effort cleanup of its storage objects.
      operationId: deleteVersion
      responses:
        "204":
          description: Version deleted
        "400":
          description: Malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Tags ------------------------------------------------------------------
  /api/v1/artifacts/{id}/tags/{tag}:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/tag"

    put:
      tags: [Tags]
      summary: Assign tag to a version
      description: >
        Atomically assigns the tag to the specified version.
        If the tag was previously assigned to another version it is moved.
      operationId: assignTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTagRequest"
      responses:
        "200":
          description: Tag assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "400":
          description: Invalid request body or malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Artifact or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Tags]
      summary: Delete tag
      operationId: deleteTag
      responses:
        "204":
          description: Tag deleted
        "404":
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Files -----------------------------------------------------------------
  /api/v1/artifacts/{id}/versions/{semver}/files:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/semver"

    get:
      tags: [Files]
      summary: List files
      operationId: listFiles
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileResponse"
        "400":
          description: Malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Artifact or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Files]
      summary: Upload file
      description: >
        Uploads a file to the version using multipart/form-data with a `file` field.
        The file name must be unique within the version. Status is set to AVAILABLE on success.
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "400":
          description: Missing file field or malformed semver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Artifact or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A file with this name already exists for this version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/versions/{semver}/files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/semver"
      - $ref: "#/components/parameters/fileId"

    get:
      tags: [Files]
      summary: Get file metadata
      operationId: getFile
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Files]
      summary: Delete file
      description: Deletes the file record and its storage object.
      operationId: deleteFile
      responses:
        "204":
          description: File deleted
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Types -----------------------------------------------------------------
  /api/v1/types:
    get:
      tags: [Types]
      summary: List types
      description: Returns all artifact types ordered by name.
      operationId: listTypes
      responses:
        "200":
          description: Type list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeResponse"

    post:
      tags: [Types]
      summary: Create type
      operationId: createType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTypeRequest"
      responses:
        "201":
          description: Type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TypeResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A type with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/types/{typeId}:
    parameters:
      - $ref: "#/components/parameters/typeId"

    get:
      tags: [Types]
      summary: Get type
      operationId: getType
      responses:
        "200":
          description: Type found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TypeResponse"
        "404":
          description: Type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [Types]
      summary: Update type
      operationId: updateType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTypeRequest"
      responses:
        "200":
          description: Type updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TypeResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A type with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Types]
      summary: Delete type
      description: Deletes the type and removes all its artifact assignments (cascade).
      operationId: deleteType
      responses:
        "204":
          description: Type deleted
        "404":
          description: Type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/types:
    parameters:
      - $ref: "#/components/parameters/artifactId"

    get:
      tags: [Types]
      summary: List artifact types
      description: Returns all types assigned to the artifact, ordered by name.
      operationId: listArtifactTypes
      responses:
        "200":
          description: Type list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TypeResponse"
        "404":
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/types/{typeId}:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/typeId"

    put:
      tags: [Types]
      summary: Assign type to artifact
      description: >
        Assigns the type to the artifact. Idempotent — assigning an already-assigned
        type returns 200 without error.
      operationId: assignArtifactType
      responses:
        "200":
          description: Type assigned (returns the type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TypeResponse"
        "404":
          description: Artifact or type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Types]
      summary: Remove type from artifact
      description: Removes the type assignment from the artifact. Returns 404 if not assigned.
      operationId: unassignArtifactType
      responses:
        "204":
          description: Type removed
        "404":
          description: Artifact not found, type not found, or type not assigned to artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/artifacts/{id}/versions/{semver}/files/{fileId}/download:
    parameters:
      - $ref: "#/components/parameters/artifactId"
      - $ref: "#/components/parameters/semver"
      - $ref: "#/components/parameters/fileId"

    get:
      tags: [Files]
      summary: Download file
      description: Streams the file content. Responds with a `Content-Disposition` header for save-as.
      operationId: downloadFile
      responses:
        "200":
          description: File stream
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="model.onnx"'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Admin -----------------------------------------------------------------
  /api/v1/admin/artifacts/empty:
    get:
      tags: [Admin]
      summary: List artifacts with no versions
      description: >
        Returns a paginated list of artifacts that have no versions and were created
        before `olderThan`. Useful for finding abandoned artifact registrations.
      operationId: adminListEmptyArtifacts
      parameters:
        - $ref: "#/components/parameters/olderThan"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated list of empty artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactPage"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Admin]
      summary: Bulk delete artifacts with no versions
      description: >
        Deletes all artifacts that have no versions and were created before `olderThan`.
        Artifacts carrying the configured protection tag are skipped and counted in `skipped`.
      operationId: adminDeleteEmptyArtifacts
      parameters:
        - $ref: "#/components/parameters/olderThan"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/versions/empty:
    get:
      tags: [Admin]
      summary: List versions with no files
      description: >
        Returns a paginated list of versions that have no files and were created
        before `olderThan`.
      operationId: adminListVersionsWithoutFiles
      parameters:
        - $ref: "#/components/parameters/olderThan"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated list of versions without files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionPage"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Admin]
      summary: Bulk delete versions with no files
      description: >
        Deletes all versions that have no files and were created before `olderThan`.
        Versions whose parent artifact carries the configured protection tag are skipped.
      operationId: adminDeleteVersionsWithoutFiles
      parameters:
        - $ref: "#/components/parameters/olderThan"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/artifacts/no-files:
    get:
      tags: [Admin]
      summary: List artifacts with versions but no files
      description: >
        Returns a paginated list of artifacts that have at least one version but
        no files on any version, and were created before `olderThan`.
      operationId: adminListArtifactsWithoutFiles
      parameters:
        - $ref: "#/components/parameters/olderThan"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated list of artifacts without files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactPage"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Admin]
      summary: Bulk delete artifacts with versions but no files
      description: >
        Deletes all artifacts (and their versions via cascade) that have at least one
        version but no files anywhere, and were created before `olderThan`.
        Artifacts carrying the configured protection tag are skipped.
      operationId: adminDeleteArtifactsWithoutFiles
      parameters:
        - $ref: "#/components/parameters/olderThan"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          description: Missing or invalid olderThan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"